+++ pkg/sink/codec/common/helper.go
@@ -14,12 +14,118 @@
 package common
 
 import (
+	"context"
+	"database/sql"
+	"fmt"
 	"math"
 
+	"github.com/go-sql-driver/mysql"
 	"github.com/pingcap/log"
 	"github.com/pingcap/tiflow/pkg/errors"
 	"go.uber.org/zap"
 )
+
+// ColumnsHolder is used to query data from sql.Rows
+type ColumnsHolder struct {
+	Values        []interface{}
+	ValuePointers []interface{}
+	Types         []*sql.ColumnType
+}
+
+func newColumnHolder(rows *sql.Rows) (*ColumnsHolder, error) {
+	columnTypes, err := rows.ColumnTypes()
+	if err != nil {
+		return nil, errors.Trace(err)
+	}
+
+	values := make([]interface{}, len(columnTypes))
+	valuePointers := make([]interface{}, len(columnTypes))
+	for i := range values {
+		valuePointers[i] = &values[i]
+	}
+
+	return &ColumnsHolder{
+		Values:        values,
+		ValuePointers: valuePointers,
+		Types:         columnTypes,
+	}, nil
+}
+
+// Length return the columns count
+func (h *ColumnsHolder) Length() int {
+	return len(h.Values)
+}
+
+// SnapshotQuery snapshot query from the given db by the commitTs and query conditions.
+func SnapshotQuery(
+	ctx context.Context, db *sql.DB, commitTs uint64, schema, table string, conditions map[string]interface{},
+) (*ColumnsHolder, error) {
+	// 1. set snapshot read
+	query := fmt.Sprintf("set @@tidb_snapshot=%d", commitTs)
+	conn, err := db.Conn(ctx)
+	if err != nil {
+		log.Error("establish connection to the upstream tidb failed",
+			zap.String("schema", schema), zap.String("table", table),
+			zap.Uint64("commitTs", commitTs), zap.Error(err))
+		return nil, errors.Trace(err)
+	}
+	defer conn.Close()
+
+	_, err = conn.ExecContext(ctx, query)
+	if err != nil {
+		mysqlErr, ok := errors.Cause(err).(*mysql.MySQLError)
+		if ok {
+			// Error 8055 (HY000): snapshot is older than GC safe point
+			if mysqlErr.Number == 8055 {
+				log.Error("set snapshot read failed, since snapshot is older than GC safe point")
+			}
+		}
+
+		log.Error("set snapshot read failed",
+			zap.String("schema", schema), zap.String("table", table),
+			zap.Uint64("commitTs", commitTs), zap.Error(err))
+		return nil, errors.Trace(err)
+	}
+
+	// 2. query the whole row
+	query = fmt.Sprintf("select * from `%s`.`%s` where ", schema, table)
+	var whereClause string
+	for name, value := range conditions {
+		if whereClause != "" {
+			whereClause += " and "
+		}
+		whereClause += fmt.Sprintf("`%s` = '%v'", name, value)
+	}
+	query += whereClause
+
+	rows, err := conn.QueryContext(ctx, query)
+	if err != nil {
+		log.Error("query row failed",
+			zap.String("schema", schema), zap.String("table", table),
+			zap.Uint64("commitTs", commitTs), zap.Error(err))
+		return nil, errors.Trace(err)
+	}
+	defer rows.Close()
+
+	holder, err := newColumnHolder(rows)
+	if err != nil {
+		log.Error("obtain the columns holder failed",
+			zap.String("schema", schema), zap.String("table", table),
+			zap.Uint64("commitTs", commitTs), zap.Error(err))
+		return nil, err
+	}
+	for rows.Next() {
+		err = rows.Scan(holder.ValuePointers...)
+		if err != nil {
+			log.Error("scan row failed",
+				zap.String("schema", schema), zap.String("table", table),
+				zap.Uint64("commitTs", commitTs), zap.Error(err))
+			return nil, errors.Trace(err)
+		}
+	}
+
+	return holder, nil
+}
 
 // BinaryLiteralToInt convert bytes into uint64,
 // by follow https://github.com/pingcap/tidb/blob/e3417913f58cdd5a136259b902bf177eaf3aa637/types/binary_literal.go#L105
